The Real McCOIL proportional method function

This function triggers the c code for the proportional method

McCOIL_proportional(
  dataA1,
  dataA2,
  maxCOI = 25,
  totalrun = 10000,
  burnin = 1000,
  M0 = 15,
  epsilon = 0.02,
  err_method = 1,
  path = getwd(),
  output = "output.txt",
  thin = 1
)

Arguments

dataA1
The intensity of signals of allele 1 from the SNP assay. Row names are names of samples and column names are names of assays.

dataA2
The intensity of signals of allele 2 from the SNP assay. Row names are names of samples and column names are names of assays.

maxCOI
Upper bound for COI. The default is 25.

totalrun
The total number of MCMC iterations. The default is 10000.

burnin
The total number of burnin iterations. The default is 1000.

M0
Initial COI. The default is 15.

epsilon
The level of measurement error (eest). The default is 0.2.

err_method
The default is 1. 1: use pre-specified epsilon; 2: use likelihood-free sampling for epsilon; 3: update epsilon according to likelihood (for 2 and 3, pre-specified epsilon was used as initial value)

path
The default is the current directory.

output
The name of output file. The default is output.txt.

thin
Numeric for how much the chain is thinned. Default = 1, means every iteration is written to file. If it was 0.1 then every 10th iteration is written to file

Value

summary of output as data.frame